chore(core) more types
