+++ modules/core/bundle.ts
@@ -2,4 +2,4 @@
 // Import from package name instead of relative path
 // This will be resolved to src or dist by esbuild depending on bundle settings
 // dist has TS transformers applied
-export * from '@deck.gl/core';
+export * from '@hubble.gl/core';
