+++ modules/core/src/animations/animation-manager.ts
@@ -10,7 +10,7 @@ export default class AnimationManager {
   timeline: Timeline;
   animations: {[id: string]: Animation} = {};
 
-  constructor({timeline = undefined, animations = []}: {timeline?: Timeline, animations: Animation[]}) {
+  constructor({timeline = undefined, animations = []}: {timeline?: Timeline, animations?: Animation[]}) {
     this.timeline = timeline || new Timeline();
     for (const animation of animations) {
       this.attachAnimation(animation);
